updated tests, handled edge case failings, force protect authorized routes
