+++ imports/components/forms/Select.js
@@ -79,6 +79,16 @@ export default class Select extends Component {
         currentTarget: target,
       });
     }
+    // if (!this.props.items.length && nextProps.items.length && nextProps.defaultValue) {
+    //   this.setValue(nextProps.defaultValue);
+    //   this.setState({ focused: false });
+    //   console.log('here');
+    //   this.change({
+    //     value: nextProps.defaultValue,
+    //     id: target.id,
+    //     currentTarget: target,
+    //   })
+    // }
   }
 
   focus = () => { // eslint-disable-line
@@ -200,7 +210,7 @@ export default class Select extends Component {
           disabled={this.disabled()}
           onFocus={this.focus}
           onChange={this.change}
-          defaultValue={this.props.items.length && this.props.defaultValue}
+          defaultValue={this.props.defaultValue}
           value={this.props.selected}
 
         >
